ChangePasswordByAdminTests added. User has his password changed by Admin
JDBC dependancy used in order to select random user from db.
On this basis the Admin changes sends the request to change password.
Mail is received on mail server using SMTP and a link to password change
is extracted.
Additionally, a method is added to transfer configuration file using FTP
protocol.